Update storage path.
This change makes Dirk pay attention to user-supplied values for the
variable `server.storage-path`, which was ignored previously.

Due to the fact that this fix could cause an issue with slashing
protection, where slashing protection was previousl written to one
location and with this fix is now written to another, the configuration
value has also moved from `server.storage-path` to just `storage-path`.
This means that slashing protection will continue to work as expected
with this change, and still allow users to change their storage path if
desired.