Add account caching to fetcher.
The memory fetcher now carries out both positive and negative account
caching.  This results in a speedup for the common cases of signging, at
the cost of no longer being able to dynamically add accounts out of
band.  Accounts added through distributed key generation are cached as
part of the generation process.